pyproject.toml
[project]
name = "hive"
version = "0.1.0"
description = "Local-first Telegram bot framework"
readme = "README.md"
authors = [
    { email = "caeochoa@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "python-telegram-bot[job-queue]>=21.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "apscheduler>=3.10,<4.0",
    "pyyaml>=6.0",
    "claude-agent-sdk>=0.0.1",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "jinja2>=3.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "httpx>=0.27",
]

[project.scripts]
hive = "hive:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
